# Increased maximum visited-list pool size to 128