--- a/lib/segment/src/index/visited_pool.rs
+++ b/lib/segment/src/index/visited_pool.rs
@@ -113,7 +113,7 @@ pub struct VisitedPool {
 impl VisitedPool {
     pub fn new() -> Self {
         VisitedPool {
-            pool: RwLock::new(vec![]),
+            pool: RwLock::new(Vec::with_capacity(*POOL_KEEP_LIMIT)),
         }
     }