We observe a significant degradation in tail latencies when there are more than 16 concurrent searches. Profiling reveals that the culprit is the visited list allocation under the exclusive lock. This PR is a proposal to increase the maximum visited-list pool size from 16 to 128. This is a reasonable maximum based on logical core count available on modern server processors suitable for the most extreme cases when a single segment is being searched by all hardware threads.

We have benchmarked this change and observe an ~80-90% reduction in 99th, 99.9th, and 99.99th percentile latencies under 27 concurrent searches. The benchmarked collection consists of 5 indexed segments of 2M points each.
Prior to the change:
```
p50: 19ms
p90: 23ms
p99: 221ms
p999: 295ms
p9999: 330ms
```
After:
```
p50: 22ms
p90: 25ms
p99: 29ms
p999: 32ms
p9999: 35ms
```
The slight increase in p50 and p90 latencies is expected as we have more concurrent segment search tasks than available cores (27 * 5 = 135 tasks vs 128 cores) which are no longer bottlenecked by visited-list allocation. If the concurrency is decreased to 24, ~3ms is shaved off of all percentiles and p50 and p90 latencies return to the original values of 19ms & 23ms.

A future improvement may be deriving the maximum pool size from the configured search thread count or using a lock-free cache which can grow to satisfy bursts in traffic and later shrink to bound auxiliary memory use.

### All Submissions:

* [X] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [X] Have you followed the guidelines in our Contributing document?
* [X] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

### Changes to Core Features:

* [X] Have you added an explanation of what your changes do and why you'd like us to include them?
* [ ] Have you successfully ran tests with your changes locally?