@generall 2023-11-06
Hi @keremnc, thanks for investigating that!

I think we should take a bit deeper into the issue and try to find a right balance for allocating those visited lists.
Right now, 128 lists for only 1M vectors would mean 128Mb of RAM just for standing by. Not all deployments can afford that.

As I can see it, there are 2 possible improvement strategies:

- link default number of visited lists to the available CPU count
- Make pool dynamic in a sense, that there should be `min` and `max` number of lists (and probably, some TTL).