@kerem-signifai 2023-11-06
Thanks for your response @generall. For your example, I think that deployments which are sized in such a way that cannot afford an extra 128 MiB of memory usage will likely not be able to push concurrent searches on the same segment to 128 because the concurrency will be limited by either network IO thread pool or search thread pool sizes, and so will not see the allocation of 128 visited-lists. 
The two improvements you listed are also the two that I mentioned in the PR description. Unfortunately, I won't have time soon to tackle either of them, both of which require a non-negligible amount of rewiring circuitry. 

Would you be OK with increasing the maximum pool size to 32, or even 48?