@generall 2023-11-06
> Unfortunately, I won't have time soon to tackle either of them, both of which require a non-negligible amount of rewiring circuitry.

No worries, we will take them over. You can expect changes to be included in the next release.

I think, pointing out the bottleneck already covers 80% of the problem, so thank you again for bringing this up to our attention!