@timvisee 2023-11-14
I've updated this to match the number of CPUs, clamped between 16 and 128:

```rust
static POOL_KEEP_LIMIT: usize = num_cpus::get().clamp(16, 128);
```

I suggest to merge like this and experiment with a dynamic pool size in a new PR.