@timvisee 2023-11-15
I did some benchmarking with this change, but find it hard to be conclusive on its effect.

#### Setup

Specs on the server I've ran this on:
- 48 CPU
- 100 GB RAM
- SSD

I've used 1 million and 16 million points with 6 to 8 segments. All vectors are in RAM, I've used the default configuration, and 100% of the vectors are indexed.

To create the collection I've used:

```rust
bfb -n 1000000
bfb -n 16000000
```

To search the collection using `bfb` on [this](https://github.com/qdrant/bfb/tree/arg-p99) branch:

```
bfb -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 48
bfb -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
```

#### Results

Here are my results. _Before_ means before the changes in this PR, _After_ means including changes in this PR. These are percentile search timings for 1 million search requests.

- 1 million points, 48 parallel searches:

  |  | Before | After | % |
  |:---|:---|:---|---:|
  | p95 | 0.009992 | 0.010039 | 0.47% |
  | p99 | 0.020969 | 0.021928 | 4.57% |
  | p999 | 0.025365 | 0.025732 | 1.45% |
  | p9999 | 0.035462 | 0.035352 | 0.31% |
  | p99999 | 0.042179 | 0.042134 | 0.11% |

- 1 million points, 24 parallel searches

  |  | Before | After | % |
  |:---|:---|:---|---:|
  | p95 | 0.005201169 | 0.0052134605 | 0.24% |
  | p99 | 0.012707645 | 0.0120774105 | 5.21% |
  | p999 | 0.020248014 | 0.0202237505 | 0.11% |
  | p9999 | 0.029463714 | 0.0286472455 | 2.85% |
  | p99999 | 0.0371282235 | 0.036175781 | 2.63% |

- 16 million points, 24 parallel searches

  |  | Before | After | % |
  |:---|:---|:---|---:|
  | p95 | 0.0055071385 | 0.0054578935 | 0.90% |
  | p99 | 0.016072051 | 0.0161296905 | -0.36% |
  | p999 | 0.02077480325 | 0.020875924 | -0.48% |
  | p9999 | 0.03170889475 | 0.031561374 | 0.47% |
  | p99999 | 0.6909968985 | 0.2739429375 | 152.24% |

I've ran most `bfb` searches multiple times for which I've taken averages. During all of these CPU utilization was about 100%.

This only seems to point to an improvement in very specific scenarios. But this may very well be within margin of error as there are a lot of other factors that can affect this. For example, this was on a virtual server. It is possible that I'm missing something important here.

In any case, I'd like to see a more well defined difference before working on a dynamic visited list pool.

<details>
<summary>Click here for <code>bfb</code> output</summary>

### 1M points, 48 parallel searches

#### Results

|  | Before | After | % |
|:---|:---|:---|---:|
| p95 | 0.009992 | 0.010039 | 0.47% |
| p99 | 0.020969 | 0.021928 | 4.57% |
| p999 | 0.025365 | 0.025732 | 1.45% |
| p9999 | 0.035462 | 0.035352 | 0.31% |
| p99999 | 0.042179 | 0.042134 | 0.11% |

#### Before PR

```bash
bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 48
--- Server timings ---
Min search time: 0.001307338
Avg search time: 0.009069664307656848
p95 search time: 0.010002134
p99 search time: 0.021268522
p999 search time: 0.025459446
p9999 search time: 0.035794997
p99999 search time: 0.043872739
Max search time: 0.05188185

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 48
--- Server timings ---
Min search time: 0.000964861
Avg search time: 0.00904691444737952
p95 search time: 0.009980915
p99 search time: 0.020669321
p999 search time: 0.025269746
p9999 search time: 0.035128486
p99999 search time: 0.04048626
Max search time: 0.048853696
```

#### After PR

```bash
$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 48
--- Server timings ---
Min search time: 0.001121642
Avg search time: 0.009135255135632297
p95 search time: 0.010091873
p99 search time: 0.022844338
p999 search time: 0.025590337
p9999 search time: 0.034523613
p99999 search time: 0.040332229
Max search time: 0.043668014

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 48
--- Server timings ---
Min search time: 0.001091625
Avg search time: 0.009029780277394294
p95 search time: 0.00998645
p99 search time: 0.021010663
p999 search time: 0.025872987
p9999 search time: 0.036180105
p99999 search time: 0.043935746
Max search time: 0.056757275
```

### 1M points, 24 parallel searches

#### Results

|  | Before | After | % |
|:---|:---|:---|---:|
| p95 | 0.005201169 | 0.0052134605 | 0.24% |
| p99 | 0.012707645 | 0.0120774105 | 5.21% |
| p999 | 0.020248014 | 0.0202237505 | 0.11% |
| p9999 | 0.029463714 | 0.0286472455 | 2.85% |
| p99999 | 0.0371282235 | 0.036175781 | 2.63% |

#### Before PR:

```bash
$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001059093
Avg search time: 0.004277637481973976
p95 search time: 0.005184454
p99 search time: 0.013752043
p999 search time: 0.020279873
p9999 search time: 0.03050666
p99999 search time: 0.038178938
Max search time: 0.052485108

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.00094141
Avg search time: 0.004305167599871859
p95 search time: 0.005217884
p99 search time: 0.011663247
p999 search time: 0.020216155
p9999 search time: 0.028420768
p99999 search time: 0.036077509
Max search time: 0.047885745
```

#### After PR:

```bash
$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001042805
Avg search time: 0.0042780262268529755
p95 search time: 0.005204907
p99 search time: 0.013321566
p999 search time: 0.020209944
p9999 search time: 0.028273203
p99999 search time: 0.036118375
Max search time: 0.060054488

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001061704
Avg search time: 0.0042833782022431354
p95 search time: 0.005222014
p99 search time: 0.010833255
p999 search time: 0.020237557
p9999 search time: 0.029021288
p99999 search time: 0.036233187
Max search time: 0.048158313
```

### 16M points, 24 parallel searches

#### Results

|  | Before | After | % |
|:---|:---|:---|---:|
| p95 | 0.0055071385 | 0.0054578935 | 0.90% |
| p99 | 0.016072051 | 0.0161296905 | -0.36% |
| p999 | 0.02077480325 | 0.020875924 | -0.48% |
| p9999 | 0.03170889475 | 0.031561374 | 0.47% |
| p99999 | 0.6909968985 | 0.2739429375 | 152.24% |

#### Before PR:

```bash
$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001794983
Avg search time: 0.0045827282695592095
p95 search time: 0.005418576
p99 search time: 0.016356783
p999 search time: 0.020757522
p9999 search time: 0.030458047
p99999 search time: 0.03989617
Max search time: 1.399457682

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001555078
Avg search time: 0.004608275713321219
p95 search time: 0.005546069
p99 search time: 0.016040071
p999 search time: 0.020713651
p9999 search time: 0.031697758
p99999 search time: 0.155489295
Max search time: 2.6827043919999998

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001557508
Avg search time: 0.004604202474350047
p95 search time: 0.005481976
p99 search time: 0.016238455
p999 search time: 0.020836185
p9999 search time: 0.032394067
p99999 search time: 0.284053025
Max search time: 0.406165842

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001625273
Avg search time: 0.0046618669314120375
p95 search time: 0.005581933
p99 search time: 0.015652895
p999 search time: 0.020791855
p9999 search time: 0.032285707
p99999 search time: 2.284549104
Max search time: 2.289780386
```

#### After PR:

```bash
$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001590478
Avg search time: 0.004541536302798149
p95 search time: 0.005453203
p99 search time: 0.015670439
p999 search time: 0.020727101
p9999 search time: 0.030903418
p99999 search time: 0.042286835
Max search time: 0.107031979

$ bfb --uri http://167.71.75.247:6334/ -n 1000000 --skip-create --skip-upload --skip-wait-index --search --p9 5 -p 24
--- Server timings ---
Min search time: 0.001447452
Avg search time: 0.004575277877977741
p95 search time: 0.005462584
p99 search time: 0.016588942
p999 search time: 0.021024747
p9999 search time: 0.03221933
p99999 search time: 0.50559904
Max search time: 0.92609554
```

</summary>